Consolated disk_writer and udp_transmit into a single set of source files to reduce code duplication.  Cleaned up the bifrost.packet_writer module to add a new _WriterBase class.
